web/requirements.txt
python-dotenv>=0.10.3
Pyrebase>=3.0.27
firebase_admin
